requirements.txt
anyio==4.2.0
appdirs==1.4.4
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
frozendict==2.3.10
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.25.2
idna==3.6
lxml==4.9.3
multitasking==0.0.11
numpy==1.26.2
pandas==2.1.4
peewee==3.17.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.7
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tzdata==2023.3
urllib3==2.1.0
webencodings==0.5.1
yfinance==0.2.33
